test support code refactoring
